chore: Update dependencies
This commit updates the following dependencies in `package.json`:
- `lucide-react` from `^0.323.0` to `^0.331.0`
- `react-router-dom` from `^6.22.0` to `^6.22.1`
- `@types/node` from `^20.11.16` to `^20.11.19`
- `@types/react` from `^18.2.55` to `^18.2.56`
- `@types/react-dom` from `^18.2.18` to `^18.2.19`
- `@typescript-eslint/eslint-plugin` from `^6.21.0` to `^7.0.1`
- `@typescript-eslint/parser` from `^6.21.0` to `^7.0.1`
- `vite` from `^5.0.12` to `^5.1.3`

Also, in `App.tsx`, the import statement for `LoginPage` was moved to the top, and a new import statement for `RoutePaths` was added.

In `useAuthentication.ts`, the `userTokenAtom` state was changed from a string to a `TokenResponse` object.

In `main.tsx`, the import statement for `BrowserRouter` was removed.

In `HomePage.tsx`, the import statement for `Link` was updated.

In `LoginPage.tsx`, an `useEffect` hook was added to redirect to the home page if a `userToken` is present.